+++ seismostats/analysis/estimate_mc.py
@@ -3,6 +3,7 @@ for the estimation of the completeness magnitude.
 """
 
 import warnings
+from typing import Any
 
 import numpy as np
 
@@ -48,6 +49,7 @@ def ks_test_gr(
     b_value: float,
     n: int = 10000,
     ks_ds: list | None = None,
+    weights: np.ndarray | None = None,
 ) -> tuple[float, float, list[float]]:
     """
     Performs the Kolmogorov-Smirnov (KS) test for the Gutenberg-Richter
@@ -67,6 +69,7 @@ def ks_test_gr(
         ks_ds:      KS distances from synthetic data with the given
                 paramters. If None, they will be estimated here (then, n is
                 not needed).
+        weights:    Array of weights for the magnitudes.
 
     Returns:
         p_val:      p-value.
@@ -86,15 +89,14 @@ def ks_test_gr(
             return 0, 1, []
 
     beta = b_value_to_beta(b_value)
+    n_sample = len(magnitudes)
 
     if ks_ds is None:
         ks_ds = []
 
-        n_sample = len(magnitudes)
-        simulated_all = simulate_magnitudes_binned(
-            n * n_sample, b_value, mc, delta_m, b_parameter="b_value"
-        )
-        max_considered_mag = np.max([np.max(magnitudes), np.max(simulated_all)])
+        # max considered magnitude: less than 1e-3 probability of
+        # exceedance within the  samples
+        max_considered_mag = 1 / b_value * np.log10(n_sample * n * 1e3) + mc
 
         x_bins = bin_to_precision(
             np.arange(mc, max_considered_mag + 3
@@ -105,13 +107,14 @@ def ks_test_gr(
         _, y_th = cdf_discrete_exp(
             x, mc=mc, delta_m=delta_m, beta=beta)
 
+        ks_ds = np.empty(n)
         for ii in range(n):
-            simulated = simulated_all[n_sample * ii: n_sample * (ii + 1)]
-            y_hist, _ = np.histogram(simulated, bins=x_bins)
+            simulated = simulate_magnitudes_binned(
+                n_sample, b_value, mc, delta_m, b_parameter="b_value"
+            )
+            y_hist, _ = np.histogram(simulated, bins=x_bins, weights=weights)
             y_emp = np.cumsum(y_hist) / np.sum(y_hist)
-
-            ks_d = np.max(np.abs(y_emp - y_th))
-            ks_ds.append(ks_d)
+            ks_ds[ii] = np.max(np.abs(y_emp - y_th))
 
     else:
         max_considered_mag = np.max(magnitudes)
@@ -123,7 +126,7 @@ def ks_test_gr(
         x_bins -= delta_m / 2
         _, y_th = cdf_discrete_exp(x, mc=mc, delta_m=delta_m, beta=beta)
 
-    y_hist, _ = np.histogram(magnitudes, bins=x_bins)
+    y_hist, _ = np.histogram(magnitudes, bins=x_bins, weights=weights)
     y_emp = np.cumsum(y_hist) / np.sum(y_hist)
 
     ks_d_obs = np.max(np.abs(y_emp - y_th))
@@ -144,8 +147,7 @@ def estimate_mc_ks(
     ks_ds_list: list[list] | None = None,
     verbose: bool = False,
     **kwargs,
-) -> tuple[float | None, float | None, list[float],
-           list[float], list[float], list[float]]:
+) -> tuple[float | None, dict[str, Any]]:
     """
     Returns the smallest magnitude in a given list of completeness magnitudes
     for which the KS test is passed, i.e., where the null hypothesis that the
@@ -186,11 +188,44 @@ def estimate_mc_ks(
 
     Returns:
         best_mc:        ``mc`` for which the p-value is lowest.
-        best_b_value:   ``b_value`` corresponding to the best ``mc``.
-        mcs_test:       Tested completeness magnitudes.
-        b_values_test:  Tested b-values.
-        ks_ds:          KS distances.
-        p_values:       Corresponding p-values.
+        mc_info:
+            Dictionary with additional information about the calculation
+            of the best ``mc``, including:
+
+            - best_b_value: ``b_value`` corresponding to the best ``mc``.
+            - mcs_tested: Tested completeness magnitudes.
+            - b_values_tested: Tested b-values.
+            - ks_ds: KS distances.
+            - p_values: Corresponding p-values.
+
+    Examples:
+        .. code-block:: python
+
+            >>> from seismostats.analysis import estimate_mc_ks
+            >>> import numpy as np
+            >>> magnitudes = np.array([2.3, 1.2, 1.5, 1.2, 1.7, 1.1, 1.2,
+            ...                   1.8, 1.6, 1.2, 1.5, 1.2, 1.7, 1.6, 1.1,
+            ...                   1.1, 1.2, 2.0, 1.1, 1.2, 1.1, 1.2, 1.6,
+            ...                   1.9, 1.3, 1.7, 1.3, 1.0, 1.2, 1.7, 1.3,
+            ...                   1.3, 1.1, 1.5, 1.4, 1.5]
+            >>> delta_m = 0.1
+            >>> mc, _ = estimate_mc_ks(magnitudes, delta_m=delta_m)
+            >>> mc
+
+            1.0
+
+        The mc_ks method returns additional information about the
+        calculation of the best mc, like b-values tested and ks
+        distances. Those are returned by the method and can be
+        used for further analysis.
+
+        .. code-block:: python
+
+            >>> best_mc, mc_info = estimate_mc_ks(magnitudes,delta_m=delta_m)
+            >>> (mc_info['b_values_tested'], mc_info['ks_ds'])
+
+            ([0.9571853220063774], [0.1700244200244202])
+
     """
 
     if mcs_test is None:
@@ -280,15 +315,20 @@ def estimate_mc_ks(
         if verbose:
             print("None of the mcs passed the test.")
 
-    return best_mc, best_b_value, mcs_tested, b_values_test, \
-        ks_ds, p_values.tolist()
+    return_vals = {'best_b_value': best_b_value,
+                   'mcs_tested': mcs_tested,
+                   'b_values_tested': b_values_test,
+                   'ks_ds': ks_ds,
+                   'p_values': p_values.tolist()}
+
+    return best_mc, return_vals
 
 
 def estimate_mc_maxc(
     magnitudes: np.ndarray,
-    delta_m: float,
+    fmd_bin: float,
     correction_factor: float = 0.2,
-) -> float:
+) -> tuple[float, dict[str, Any]]:
     """
     Returns the completeness magnitude (mc) estimate using the maximum
     curvature method.
@@ -305,8 +345,8 @@ def estimate_mc_maxc(
 
     Args:
         magnitudes:         Array of magnitudes to test.
-        delta_m:            Bin size for the maximum curvature method. This can
-                        be independent ofthe descritization of the magnitudes.
+        fmd_bin:            Bin size for the maximum curvature method. This can
+                        be independent of the discretization of the magnitudes.
                         The original value for the maximum curvature method is
                         0.1. However, the user can decide which value to use.
                         The optimal value would be as small as possible while
@@ -314,16 +354,49 @@ def estimate_mc_maxc(
                         magnitudes in each bin. If the bin size is too small,
                         the method will not work properly.
         correction_factor:  Correction factor for the maximum curvature
-                method (default value after Woessner & Wiemer 2005).
+                        method (default value +0.2 after Woessner &
+                        Wiemer 2005).
 
     Returns:
         mc:                 Estimated completeness magnitude.
+        mc_info:
+            Dictionary with additional information about the calculation
+            of the best ``mc``, including:
+
+            - correction_factor:   Correction factor for the maximum curvature
+              method (default value +0.2 after Woessner & Wiemer 2005).
+
+
+    Examples:
+        .. code-block:: python
+
+            >>> from seismostats.analysis import estimate_mc_maxc
+            >>> import numpy as np
+            >>> magnitudes = np.array([2.3, 1.2, 1.5, 1.2, 1.7, 1.1, 1.2,
+            ...                   1.8, 1.6, 1.2, 1.5, 1.2, 1.7, 1.6, 1.1,
+            ...                   1.1, 1.2, 2.0, 1.1, 1.2, 1.1, 1.2, 1.6,
+            ...                   1.9, 1.3, 1.7, 1.3, 1.0, 1.2, 1.7, 1.3,
+            ...                   1.3, 1.1, 1.5, 1.4, 1.5]
+            >>> delta_m = 0.1
+            >>> mc, _ = estimate_mc_maxc(magnitudes, delta_m=delta_m)
+            >>> mc
+            1.4
+
+        The mc_maxc method also returns the correction factor used in the
+        calculation of the best mc value.
+
+        .. code-block:: python
+
+            >>> best_mc, mc_info = cat.estimate_mc_maxc(fmd_bin=0.1)
+            >>> mc_info['correction_factor']
+            0.2
+
     """
     bins, count, _ = get_fmd(
-        magnitudes=magnitudes, delta_m=delta_m, bin_position="center"
+        magnitudes=magnitudes, fmd_bin=fmd_bin, bin_position="center"
     )
     mc = bins[count.argmax()] + correction_factor
-    return mc
+    return mc, {'correction_factor': correction_factor}
 
 
 def estimate_mc_b_stability(
@@ -335,8 +408,7 @@ def estimate_mc_b_stability(
         stability_range: float = 0.5,
         verbose: bool = False,
         **kwargs,
-) -> tuple[float | None, float | None, list[float],
-           list[float], list[float], list[float]]:
+) -> tuple[float | None, dict[str, Any]]:
     """
     Estimates the completeness magnitude (mc) using b-value stability.
 
@@ -345,10 +417,13 @@ def estimate_mc_b_stability(
     for the stability test by changing the stability_range.
 
     Source:
-        Woessner, J, and Stefan W. "Assessing the quality of earthquake
-        catalogues: Estimating the magnitude of completeness and its
-        uncertainty." Bulletin of the Seismological Society of America 95.2
-        (2005): 684-698.
+        - Cao, A., & Gao, S. S. (2002). Temporal variation of seismic b-values
+            beneath northeastern Japan island arc. Geophysical Research Letters,
+            29(9), 1–3. https://doi.org/10.1029/2001gl013775
+        - Woessner, J, and Stefan W. "Assessing the quality of earthquake
+            catalogues: Estimating the magnitude of completeness and its
+            uncertainty." Bulletin of the Seismological Society of America 95.2
+            (2005): 684-698.
 
     Args:
         magnitudes:         Array of magnitudes.
@@ -370,12 +445,49 @@ def estimate_mc_b_stability(
 
     Returns:
         best_mc:        Best magnitude of completeness estimate.
-        best_b_value:   b-value associated with ``best_mc``.
-        mcs_test:       Array of tested completeness magnitudes.
-        b_values_test:  Array of b-values associated to tested mcs.
-        diff_bs:        Array of differences divided by std, associated
-                    with tested mcs. If a value is smaller than one, this
-                    means that the stability criterion is met.
+        mc_info:
+            Dictionary with additional information about the calculation
+            of the best ``mc``, including:
+
+            - best_b_value: b-value associated with ``best_mc``.
+            - mcs_tested:     Array of tested completeness magnitudes.
+            - b_values_tested: Array of b-values associated to tested mcs.
+            - diff_bs:  Array of differences divided by std, associated
+              with tested mcs. If a value is smaller than one,
+              this means that the stability criterion is met.
+
+    Examples:
+        .. code-block:: python
+
+            >>> from seismostats.analysis import estimate_mc_b_stability
+            >>> import numpy as np
+            >>> magnitudes = np.array([2.3, 1.2, 1.5, 1.2, 1.7, 1.1, 1.2,
+            ...                   1.8, 1.6, 1.2, 1.5, 1.2, 1.7, 1.6, 1.1,
+            ...                   1.1, 1.2, 2.0, 1.1, 1.2, 1.1, 1.2, 1.6,
+            ...                   1.9, 1.3, 1.7, 1.3, 1.0, 1.2, 1.7, 1.3,
+            ...                   1.3, 1.1, 1.5, 1.4, 1.5]
+            >>> delta_m = 0.1
+            >>> mc, _ = estimate_mc_b_stability(magntitudes,delta_m=delta_m)
+            >>> mc
+
+            1.1
+
+        The mc_b_stability method returns additional information about the
+        calculation of the best mc, like b-values tested and the array of
+        differences. Those are returned by the method and can be used for
+        further analysis.
+
+        .. code-block:: python
+
+            >>> best_mc, mc_info = estimate_mc_b_stability(magnitudes,
+            ...     delta_m=delta_m)
+            >>> (mc_info['b_values_tested'], mc_info['diff_bs'])
+
+            ([np.float64(0.9571853220063772),
+            np.float64(1.190298769977797)],
+            [np.float64(2.2337527711215786),
+            np.float64(0.9457747650207581)])
+
     """
     steps = len(np.arange(0, stability_range, delta_m))
 
@@ -454,5 +566,12 @@ def estimate_mc_b_stability(
     elif not value:
         print("None of the mcs passed the stability test.")
 
-    return bin_to_precision(best_mc, delta_m) if best_mc else None, \
-        best_b_value, mcs_test.tolist(), b_values_test, diff_bs
+    return_vals = {
+        'best_b_value': best_b_value,
+        'mcs_tested': mcs_test.tolist(),
+        'b_values_tested': b_values_test,
+        'diff_bs': diff_bs
+    }
+    print(f'return_vals: {return_vals}')
+    return (bin_to_precision(best_mc, delta_m) if best_mc else None,
+            return_vals)